refactor: 커스텀 텍스트필드 이름 변경 및 패키지 이동
