+++ core/designsystem/src/main/java/com/saeongjima/designsystem/component/textfield/MainTextField.kt
@@ -1,4 +1,4 @@
-package com.saeongjima.designsystem.component
+package com.saeongjima.designsystem.component.textfield
 
 import androidx.annotation.DrawableRes
 import androidx.compose.animation.animateColor
