+++ feature/onboading/src/main/java/com/saeongjima/signup/PersonalInformationScreen.kt
@@ -25,7 +25,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.saeongjima.designsystem.component.DanjamTextField
+import com.saeongjima.designsystem.component.textfield.DanjamTextField
 import com.saeongjima.designsystem.component.button.DanjamButton
 import com.saeongjima.designsystem.theme.Black100
 import com.saeongjima.designsystem.theme.Black300
